Fix pool statistics underflow bug
Bug: current_allocated counter was showing garbage values (underflow)

Root cause:
1. Pre-allocated buffers incremented current_allocated (wrong - they're free)
2. Buffers from free list didn't increment current_allocated
3. All returns decremented current_allocated
4. This caused underflow when returning pre-allocated buffers

Fix:
1. Constructor: Don't increment current_allocated for pre-allocated buffers
2. allocate(): Always increment current_allocated when handing out buffer
3. returnBuffer(): Unchanged - decrements on return

Also fix format specifiers: use %llu for uint64_t values

All 15 tests passing.

+++ main.cpp
@@ -297,10 +297,13 @@ int main(int argc, char *argv[]) {
         auto pool_stats = client.getPoolStats();
         LOG_INFO("[PoolStats] === Memory Pool Statistics ===");
         for (const auto& s : pool_stats) {
-          LOG_INFO("[PoolStats] Size=%zuB allocated=%lu free=%zu in_use=%lu "
-                   "misses=%lu",
-                   s.block_size, s.total_allocated, s.free_count,
-                   s.current_allocated, s.pool_misses);
+          LOG_INFO("[PoolStats] Size=%zuB allocated=%llu free=%zu in_use=%llu "
+                   "misses=%llu",
+                   s.block_size, 
+                   (unsigned long long)s.total_allocated, 
+                   s.free_count,
+                   (unsigned long long)s.current_allocated, 
+                   (unsigned long long)s.pool_misses);
         }
         last_pool_print = now;
       }
